Rewrite version progress 2:
- Move every parser to it's own version
- Delete parser.py, move the get_parser function to common.py
- Add better error handling.
- Add information about this branch

- Finalized GYAO, will do some testing later.
- Add ponimu.py (Contain some information)
- Add bilibili.py (Still empty)
