+++ yuu/command.py
@@ -5,9 +5,8 @@ import subprocess
 import click
 import requests
 
-from .common import __version__, isUserAdmin, res_data
+from .common import __version__, isUserAdmin, get_parser
 from .downloader import download_chunk, merge_video
-from .parser import get_parser
 
 CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'], ignore_unknown_options=True)
 
@@ -19,7 +18,7 @@ def cli(version=False, update=False):
     A simple AbemaTV video downloader
     """
     if version:
-        print('yuu v{} - Created by NoAiOne'.format(__version__))
+        print('yuu rewrite v{} - Created by NoAiOne'.format(__version__))
         exit(0)
     if update:
         print('[INFO] Updating yuu...'.format(ver=__version__))
@@ -51,33 +50,18 @@ def cli(version=False, update=False):
         exit(0)
 
 
-# @cli.command("download", short_help="Download video from abema.tv")
-# @click.argument("input", metavar="<AbemaTV url site or m3u8>")
-# @click.option("--proxy", "-p", required=False, default=None, metavar="<ip:port/url>", help="Use http(s)/socks5 proxies (please add `socks5://` if you use socks5)")
-# @click.option("--resolution", "-r", "res", required=False, type=click.Choice(['180p', '240p', '360p', '480p', '720p', '1080p']), default="1080p", help="Resolution to be downloaded (Default: 1080p)")
-# @click.option("--resolutions", "-R", "resR", is_flag=True, help="Show available resolutions")
-# @click.option("--output", "-o", required=False, default=None, help="Output filename")
-# @click.option('--verbose', '-v', is_flag=True, help="Enable verbosity")
-# def main_downloaderv2(input, proxy, res, resR, output, verbose):
-#     print('[INFO] Starting yuu v{ver}...'.format(ver=__version__))
-
-#     upstream_data = requests.get("https://pastebin.com/raw/Bt3ZLjfu").json()
-#     upstream_version = upstream_data['version']
-#     if upstream_version != __version__:
-#         print('[INFO] There\'s new version available to download, please update using `yuu -U`.')
-#         exit(0)
-
-
-@cli.command("download", short_help="Download video from abema.tv")
-@click.argument("input", metavar="<AbemaTV url site or m3u8>")
+@cli.command("download", short_help="Download video from abema.tv or gyao")
+@click.argument("input", metavar="<URL site>")
+@click.option("--username", "-U", required=False, default=None, help="Use username/password to download premium video")
+@click.option("--password", "-P", required=False, default=None, help="Use username/password to download premium video")
 @click.option("--proxy", "-p", required=False, default=None, metavar="<ip:port/url>", help="Use http(s)/socks5 proxies (please add `socks5://` if you use socks5)")
 @click.option("--resolution", "-r", "res", required=False, type=click.Choice(['180p', '240p', '360p', '480p', '720p', '1080p']), default="1080p", help="Resolution to be downloaded (Default: 1080p)")
 @click.option("--resolutions", "-R", "resR", is_flag=True, help="Show available resolutions")
 @click.option("--output", "-o", required=False, default=None, help="Output filename")
 @click.option('--verbose', '-v', is_flag=True, help="Enable verbosity")
-def main_downloader(input, proxy, res, resR, output, verbose):
-    """Download a free video from abema"""
-    print('[INFO] Starting yuu v{ver}...'.format(ver=__version__))
+def main_downloader(input, username, password, proxy, res, resR, output, verbose):
+    """Main command to access downloader"""
+    print('[INFO] Starting yuu rewrite v{ver}...'.format(ver=__version__))
 
     #upstream_data = requests.get("https://pastebin.com/raw/Bt3ZLjfu").json()
     #upstream_version = upstream_data['version']
@@ -106,7 +90,6 @@ def main_downloader(input, proxy, res, resR, output, verbose):
                 if mode == proxy_test[-1]:
                     print('[ERROR] Cannot connect to proxy (Request timeout)')
                     exit(1)
-        sesi.proxies = {'http': proxy, 'https': proxy}
     try:
         sesi.get('http://httpbin.org/get')
         pmode = "No proxy"
@@ -130,10 +113,16 @@ def main_downloader(input, proxy, res, resR, output, verbose):
         mode = 'GYAO'
 
     print('[INFO] {}: Fetching user token'.format(mode))
-    yuuParser.get_token()
+    result, reason = yuuParser.get_token()
+    if not result:
+        print('[ERROR] {}: {}'.format(mode, reason))
+        exit(1)
 
     print('[INFO] {}: Parsing url'.format(mode))
-    output_name = yuuParser.parse(input, res)
+    output_name, reason = yuuParser.parse(input, res)
+    if not output_name:
+        print('[ERROR] {}: {}'.format(mode, reason))
+        exit(1)
     if resR:
         print('[INFO] {}: Checking available resolution'.format(mode))
         avares = yuuParser.resolutions()
@@ -141,7 +130,7 @@ def main_downloader(input, proxy, res, resR, output, verbose):
         print('{0: <{width}}{1: <{width}}{2: <{width}}{3: <{width}}'.format("", "Resolution", "Video Quality", "Audio Quality", width=16))
         for res in avares:
             r_c, wxh = res
-            vidq, audq = res_data[r_c]
+            vidq, audq = yuuParser.resolution_data[r_c]
             print('{0: <{width}}{1: <{width}}{2: <{width}}{3: <{width}}'.format('>> ' + r_c, wxh, vidq, audq, width=16))
         exit(0)
 
@@ -151,19 +140,31 @@ def main_downloader(input, proxy, res, resR, output, verbose):
         else:
             output = output + '.ts'
     else:
-        output = '{x} (AbemaTV {r}).ts'.format(x=output_name, r=res)
+        output = '{x} ({m} {r}).ts'.format(x=output_name, m=mode, r=res)
 
     print('[INFO] {}: Parsing m3u8'.format(mode))
-    files, iv = yuuParser.parse_m3u8()
+    files, iv, reason = yuuParser.parse_m3u8()
+
+    if not files:
+        print('[ERROR] {}: {}'.format(mode, reason))
+        exit(1)
 
     illegalchar = ['/', '<', '>', ':', '"', '\\', '|', '?', '*'] # https://docs.microsoft.com/en-us/windows/desktop/FileIO/naming-a-file
     for char in illegalchar:
         output = output.replace(char, '_')
 
     print('[INFO] {}: Fetching video key'.format(mode))
-    video_key = yuuParser.get_video_key()#fetch_video_key(ticket, authtoken, sesi, verbose)
+    video_key, reason = yuuParser.get_video_key() #fetch_video_key(ticket, authtoken, sesi, verbose)
+    if not video_key:
+        if reason != 'No Encryption': # If the returned reason are 'No Encryption' it will bypass.
+            print('[ERROR] {}: {}'.format(mode, reason))
+            exit(1)
 
     print('[INFO][DOWN] Starting downloader...')
+    if mode == 'GYAO': # To make download faster?
+        yuuParser.session.proxies = {'http': None, 'https': None}
+        print('[INFO] GYAO: If you\'re using vpn or proxy, you can disable it now\nIf you\'re using `-p` for proxy, it will be remove')
+        input('[INFO] GYAO: Press enter to continue...')
     print('[INFO][DOWN] Output: {}'.format(output))
     print('[INFO][DOWN] Resolution: {}'.format(res))
     dl_list, temp_dir = download_chunk(files, video_key, iv, yuuParser.session)
