+++ yuu/common.py
@@ -2,25 +2,9 @@ import os
 import json
 import re
 
+from .ext import *
+
 __version__ = "1.0.0"
-
-_STRTABLE = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz"
-_HKEY = b"3AF0298C219469522A313570E8583005A642E73EDD58E3EA2FB7339D3DF1597E"
-
-_KEYPARAMS = {
-    "osName": "android",
-    "osVersion": "6.0.1",
-    "osLand": "ja_JP",
-    "osTimezone": "Asia/Tokyo",
-    "appId": "tv.abema",
-    "appVersion": "3.27.1"
-}
-
-_MEDIATOKEN_API = "https://api.abema.io/v1/media/token"
-_LICENSE_API = "https://license.abema.io/abematv-hls"
-_USERAPI = "https://api.abema.io/v1/users"
-_PROGRAMAPI = 'https://api.abema.io/v1/video/programs/'
-_CHANNELAPI = 'https://api.abema.io/v1/media/slots/'
 
 def is_channel(url):
     url = re.findall('(slot)', url)
@@ -29,27 +13,15 @@ def is_channel(url):
     return False
 
 
-abema_data = {
-    "1080p": ["4000kb/s", "AAC 192kb/s 2ch"],
-    "720p": ["2000kb/s", "AAC 160kb/s 2ch"],
-    "480p": ["900kb/s", "AAC 128kb/s 2ch"],
-    "360p": ["550kb/s", "AAC 128kb/s 2ch"],
-    "240p": ["240kb/s", "AAC 64kb/s 1ch"],
-    "180p": ["120kb/s", "AAC 64kb/s 1ch"]
-}
+def get_parser(url):
+    valid_abema = r'https?://(?:abema\.tv)/(?:channels|video)/(?:\w*)(?:/|-\w*/)((?P<slot>slots/)|)(?P<video_id>.*[^-_])'
+    valid_gyao = r'(?isx)http(?:|s)://gyao.yahoo.co.jp/(?:player|title[\w])/(?P<p1>[\w]*.*)'  
+    if re.match(valid_abema, url):
+        return AbemaTV
+    elif re.match(valid_gyao, url):
+        return GYAO
+    return None
 
-abema_data = {
-    "1080p-0": ["~5000kb/s", "AAC 64kb/s 2ch"],
-    "720p-0": ["2000kb/s", "AAC 64kb/s 2ch"],
-    "480p-0": ["900kb/s", "AAC 64kb/s 2ch"],
-    "360p-0": ["550kb/s", "AAC 64kb/s 2ch"],
-    "240p-0": ["~200kb/s", "AAC 64kb/s 1ch"],
-    "1080p-1": ["~5000kb/s", "AAC 128kb/s 2ch"],
-    "720p-1": ["~2000kb/s", "AAC 128kb/s 2ch"],
-    "480p-1": ["~900kb/s", "AAC 128kb/s 2ch"],
-    "360p-1": ["~550kb/s", "AAC 128kb/s 2ch"],
-    "240p-1": ["~200kb/s", "AAC 128kb/s 2ch"],
-}
 
 """
 Admin Check Code from https://gist.github.com/sylvainpelissier/ff072a6759082590a4fe8f7e070a4952
