+++ yuu/ext/__init__.py
@@ -1,0 +1,6 @@
+"""
+Initiate all extension that are available
+"""
+
+from .abematv import AbemaTV
+from .gyao import GYAO
